Fixing concurrency issue in Memory Cache
I'm seeing some System.IndexOutOfRangeException exceptions in our
application caused by corruption of the child regions HashSet. This
appears to be caused by concurrent writes to the data structure. I'm
replacing the HashSet with a ConcurrentDictionary. This isn't ideal as
each key also has to store a bogus value, but there is no builtin
concurrent HashSet and this seemed to be a better solution than explicit
locking.